Remove hard-filtering TODO
It seems inefficient to decide to forward a packet based on whether it is
hard-filtered or not, considering there can be multiple new subscriptions in a
given packet. The forwarding can be seen as sufficiently cheap that we might as
well just pass it on rather than expend the extra time checking each contained
subscription to see if it's hard-filtered.